backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
httpx==0.27.0
pydantic==2.6.0
sentence-transformers==2.7.0
langsmith==0.1.0
langchain==0.1.0
python-dotenv==1.0.0
numpy==1.26.0
scikit-learn==1.4.0
